+++ movie/infra/src/main/kotlin/com/example/movie/entity/MovieEntity.kt
@@ -1,4 +1,4 @@
-package com.example.movie.persistence.movie.entity
+package com.example.movie.entity
 
 import com.example.movie.domain.movie.model.Movie
 import com.example.movie.domain.movie.model.Rating
