+++ movie/infra/src/main/kotlin/com/example/movie/repository/GenreJpaRepository.kt
@@ -1,6 +1,6 @@
-package com.example.movie.persistence.movie.repository
+package com.example.movie.repository
 
-import com.example.movie.persistence.movie.entity.GenreEntity
+import com.example.movie.entity.GenreEntity
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
