+++ movie/infra/src/main/kotlin/com/example/movie/repository/GenreRepositoryImpl.kt
@@ -1,4 +1,4 @@
-package com.example.movie.persistence.movie.repository
+package com.example.movie.repository
 
 import com.example.movie.domain.movie.model.Genre
 import com.example.movie.domain.movie.repository.GenreRepository
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository
 
 @Repository
 class GenreRepositoryImpl(
-    private val genreJpaRepository: com.example.movie.persistence.movie.repository.GenreJpaRepository
+    private val genreJpaRepository: GenreJpaRepository
 ) : GenreRepository {
     override fun findById(id: Long): Genre? {
         return genreJpaRepository.findByIdOrNull(id)?.toDomain()
