+++ movie/infra/src/main/kotlin/com/example/movie/repository/MovieRepositoryImpl.kt
@@ -1,4 +1,4 @@
-package com.example.movie.persistence.movie.repository
+package com.example.movie.repository
 
 import com.example.movie.domain.movie.model.Movie
 import com.example.movie.domain.movie.repository.MovieRepository
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository
 
 @Repository
 class MovieRepositoryImpl(
-    private val movieJpaRepository: com.example.movie.persistence.movie.repository.MovieJpaRepository
+    private val movieJpaRepository: MovieJpaRepository
 ) : MovieRepository {
     override fun findById(id: Long): Movie? {
         return movieJpaRepository.findById(id).orElse(null)?.toDomain()
